chore(editor): move paste.ts under editor/ (consistent module layout)
